+++ lua/snippets/luasnip/lua_snippets.lua
@@ -54,7 +54,53 @@ local prequire = s(
 	)
 )
 
+local m_snip = s(
+	{ trig = 'lm', name = 'local M', dscr = 'local M' },
+	fmta(
+		[[
+	local M = {}
+	
+	<>
+
+	return M
+	]],
+		{ i(0) }
+	)
+)
+
+local meth = s(
+	{ trig = 'meth', name = 'Method', dscr = 'Method declaration' },
+	fmta(
+		[[
+	function <receiver>.<name>(<params>)
+		<e>
+	end
+	]],
+		{
+			receiver = i(1, 'receiver'),
+			name = i(2, 'name'),
+			params = i(3),
+			e = i(0),
+		}
+	)
+)
+
+local local_f = s(
+	{ trig = 'lf', name = 'local function', dscr = 'local function declaration' },
+	fmta(
+		[[
+		local function <>(<>)
+			<>
+		end
+	]],
+		{ i(1, 'name'), i(2), i(0) }
+	)
+)
+
 ls.add_snippets('lua', {
 	packmod,
 	prequire,
+	m_snip,
+	meth,
+	local_f,
 })