+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 <a name="unreleased"></a>
 ### [Unreleased]
 
+> Code Refactoring
+- **snippet:** readable neogen / luasnip jump handling
+
 > Features
 - **snippet:** added prequire snippet lua
 - **snippet:** more snippet for go