+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - **mappings:** moved q to <leader>q
 - **snippet:** added prequire snippet lua
 - **snippet:** more snippet for go
+- **snippets:** added more lua snippets
 
 
 <a name="v3.1.6"></a>