+++ CHANGELOG.md
@@ -11,6 +11,9 @@
 - **snippets:** added dynamic preq snippet
 - **snippets:** added more lua snippets
 
+> Various Actions
+- removed print
+
 
 <a name="v3.1.6"></a>
 ### [v3.1.6] - 2022-03-31