+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 > Features
 - **mappings:** moved q to <leader>q
+- **snippet:** added dynamic req snippet
 - **snippet:** added prequire snippet lua
 - **snippet:** more snippet for go
 - **snippets:** added dynamic preq snippet