+++ CHANGELOG.md
@@ -12,6 +12,9 @@
 - **snippets:** added dynamic preq snippet
 - **snippets:** added more lua snippets
 
+> Revert
+- **snippet:** removed apm_span from auto. prefer to use <c-j> expansion instead
+
 > Various Actions
 - removed print