+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - **mappings:** moved q to <leader>q
 - **snippet:** added table generating snippets markdown lua
 - **snippet:** added dynamic req snippet
+- **snippet:** added description to make snippet golang
 - **snippet:** added prequire snippet lua
 - **snippet:** more snippet for go
 - **snippets:** added dynamic preq snippet